Rename IOTAChargingStationWiFi to IOTAChargingStationWifi
An extension of the IOTAChargingStation project using an ESP8266 WiFi modules chip instead of the Raspberry pi 3 b+ to receive the transmitted http request after approval of the transaction on the IOTA tangle. IOTA transaction confirmation is outsourced and executed to an external machine (in this case my laptop) which upon transaction validity confirmation sends an http request to the ESP8266 chip that then turns on power for the device. The above code represents the Python file used on the external machine. This project follows the tutorial by Hugo Gregersen posted on Medium and his GitHub page.